requirements.txt
ollama
python-dotenv
discord.py
